Tests no longer target the TDS at remotetest.unidata.ucar.edu * Instead, they target the embedded TDS that's launched in :it:test. Some test classes needed to be moved into :it to make that possible. Had to move around some datasets and tweak some paths as well to keep tests passing. * TestSSH, TestTomcatAuth, and TestByteRange were improved to allow execution within Jenkins and Travis. Previously, they had to be run manually from IntelliJ. * Removed NeedsExternalResource annotation from every test I could. More pruning to come in the future. * Added NeedsCdmUnitTest to a couple of classes that (sneakily) require it. * :it:appRun no longer requires :it:processTestResources to be run first.
