+++ dap4/d4tests/src/test/java/dap4/test/TestSerial.java
@@ -4,10 +4,8 @@ import dap4.core.util.DapUtil;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
 import ucar.nc2.dataset.NetcdfDataset;
 import ucar.unidata.util.test.TestDir;
-import ucar.unidata.util.test.category.NeedsExternalResource;
 
 import java.io.IOException;
 import java.io.StringWriter;
@@ -148,7 +146,6 @@ public class TestSerial extends DapTestCommon
     //////////////////////////////////////////////////
     // Junit test method
 
-    @Category(NeedsExternalResource.class)
     @Test
     public void testSerial()
         throws Exception
