+++ it/src/test/java/ucar/nc2/TestByteRange.java
@@ -34,11 +34,7 @@ package ucar.nc2;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
 import ucar.unidata.util.test.UnitTestCommon;
-import ucar.unidata.util.test.category.NeedsExternalResource;
-import ucar.unidata.util.test.category.NotJenkins;
-import ucar.unidata.util.test.category.NotTravis;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -48,7 +44,6 @@ import java.util.List;
 /**
  * Test that client side and server side byte range access works.
  */
-@Category({NotJenkins.class,NotTravis.class}) // Temporary??
 public class TestByteRange extends UnitTestCommon
 {
     // Collect testcases locally
@@ -87,7 +82,6 @@ public class TestByteRange extends UnitTestCommon
     }
 
     @Test
-    @Category(NeedsExternalResource.class)
     public void
     testByteRange() throws Exception
     {
