+++ it/src/test/java/ucar/nc2/util/net/Login.java
@@ -51,18 +51,18 @@ public class Login extends JFrame implements CredentialsProvider
     protected String dfaltuser = null;
     protected String dfaltpwd = null;
 
-  // In order to act as a proper CredentialsProvider, this class
-  // needs to support the setCredentials operation properly.
-  // This means that it must provide an internal database of
-  // AuthScope->Credentials.
-  // Because this class is interactive, entries in the internal database
-  // must only be set by explicit calls to setCredentials().
-  // Why do we need this? Because this provider will be used everywhere
-  // in the path to the server to provide credentials at intermediate point.
-  // So, it must be possible to set proxy credentials as well as 
-  // possible redirections (e.g. OAUTH2).
+    // In order to act as a proper CredentialsProvider, this class
+    // needs to support the setCredentials operation properly.
+    // This means that it must provide an internal database of
+    // AuthScope->Credentials.
+    // Because this class is interactive, entries in the internal database
+    // must only be set by explicit calls to setCredentials().
+    // Why do we need this? Because this provider will be used everywhere
+    // in the path to the server to provide credentials at intermediate point.
+    // So, it must be possible to set proxy credentials as well as
+    // possible redirections (e.g. OAUTH2).
 
-  private Map<AuthScope, Credentials> creds = new HashMap<>();
+    private Map<AuthScope, Credentials> creds = new HashMap<>();
 
 
     /**
@@ -78,24 +78,24 @@ public class Login extends JFrame implements CredentialsProvider
 
     public void clear()
     {
-      this.creds.clear();
+        this.creds.clear();
     }
 
     public void setCredentials(AuthScope scope, Credentials cred)
     {
-      this.creds.put(scope, cred);
+        this.creds.put(scope, cred);
     }
 
     public Credentials getCredentials(AuthScope scope)
     {
-     Credentials credentials = this.creds.get(scope);
-     if(credentials == null) {
-       AuthScope bestMatch = HTTPAuthUtil.bestmatch(scope,this.creds.keySet());
-       if(bestMatch != null)
-         credentials = this.creds.get(bestMatch);
-    }
-    if(credentials != null)
-      return credentials;
+        Credentials credentials = this.creds.get(scope);
+        if(credentials == null) {
+            AuthScope bestMatch = HTTPAuthUtil.bestmatch(scope,this.creds.keySet());
+            if(bestMatch != null)
+                credentials = this.creds.get(bestMatch);
+        }
+        if(credentials != null)
+            return credentials;
 
         String up = login(this.dfaltuser, this.dfaltpwd, scope);
         if(up == null) throw new IllegalStateException();
@@ -148,14 +148,14 @@ public class Login extends JFrame implements CredentialsProvider
             int optionindex = JOptionPane.showConfirmDialog(null, userPanel, title.toString(),
                     JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
             switch (optionindex) {
-            case 0: // ok
-                user = userFld.getText();
-                pwd = passwordFld.getText();
-                if(user != null && pwd != null)
-                    return user + ":" + pwd;
-                break;
-            default: // treat rest as cancel
-                break;
+                case 0: // ok
+                    user = userFld.getText();
+                    pwd = passwordFld.getText();
+                    if(user != null && pwd != null)
+                        return user + ":" + pwd;
+                    break;
+                default: // treat rest as cancel
+                    break;
             }
             break;
         }
