+++ tds/src/test/content/thredds/public/testdata/mydata1.nc
